+++ lib/src/models/models.dart
@@ -4,3 +4,4 @@ export 'call_ring_params.dart';
 export 'make_call_response.dart';
 export 'jwt_payload.dart';
 export 'ws_message.dart';
+export 'phone_number.dart';